+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed file permssions on Linux
 - Reworked file/folder/remote URL fetching dialogs
 - Fixed "jumpy" scroll bars while changing channels or messages
+- Fixed issue with some messages not being tappable 
 
 ## [0.0.2] - 2020-07-31
 ### Added